Separates baseHandler from SpeedCDF
