+++ app/handlers/baseHandler.py
@@ -1,0 +1,23 @@
+import tornado.web
+import json
+import traceback
+
+class BaseHandler(tornado.web.RequestHandler):
+    def write_error(self, status_code, **kwargs):
+        self.set_header('Content-Type', 'application/json')
+        if self.settings.get('serve_traceback') and 'exc_info' in kwargs:
+            lines = []
+            for line in traceback.format_exception(*kwargs['exc_info']):
+                lines.append(line)
+            self.finish(json.dumps({
+                'error': {
+                    'code': status_code,
+                    'message': self._reason,
+                    'traceback': lines,
+                }}))
+        else:
+            self.finish(json.dumps({
+                'error': {
+                    'code': status_code,
+                    'message': self._reason,
+                }}))
